### 2021-08-20 #3638 @whedon
[PRE REVIEW]: BCImat: a Matlab-based framework for Intracortical Brain-Computer Interfaces and their simulation with an artificial spiking neural network
**Submitting author:** @eferrea (<a href="http://orcid.org/0000-0003-1352-4789">Enrico Ferrea</a>)
**Repository:** <a href="https://github.com/eferrea/BCI-Master" target ="_blank">https://github.com/eferrea/BCI-Master</a>
**Version:** v.1.0.0
