### 2021-08-20 #3638 @kthyng
Hi  and thanks for your submission. I have pinged the greater editorial board to make sure this submission is in scope, particularly in terms of [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort). This process usually takes 1-2 weeks; thanks for your patience.
