### 2021-08-24 #3638 @eferrea
@kthyng thanks for your answer. Just few words: this work was started in the lab back to 2015 and was updated among the years to perform state of the art brain computer interface experiments for motor control. It also contains implementations of several BCI functionalities the details of which can be hardly found in papers. It is widely used from people in our lab and we hope that other people can benefit from it.
