### 2021-09-03 #3638 @kyleniemeyer
@eferrea could you also recommend some potential reviewers? Ideally from our [reviewer list](https://bit.ly/joss-reviewers) but also potentially domain experts you are aware of. Thanks!
