### 2021-09-05 #3638 @eferrea
Hi @kyleniemeyer, I added the LICENSE file to the repository. Additional potential reviewers are Agricolab, TomDonoghue and JanCBrammer.
