### 2021-09-28 #3638 @TomDonoghue
Thanks for the invite! This project looks interesting, but I don't use Matlab anymore (don't have a license / wouldn't be able to test the code), so I won't be able to do a proper review here.
