### 2021-09-29 #3638 @JanCBrammer
Thanks for the request! Unfortunately, I'm too swamped atm to properly review this. Also don't have a Matlab licence anymore and it's been quite a while since I did anything with it (will take Matlab off my language list in the reviewer sheet 🙈 ).
