### 2021-11-29 #3956 @whedon
[REVIEW]: BCImat: a Matlab-based framework for Intracortical Brain-Computer Interfaces and their simulation with an artificial spiking neural network
**Submitting author:** @eferrea (<a href="http://orcid.org/0000-0003-1352-4789">Enrico Ferrea</a>)
**Repository:** https://github.com/eferrea/BCI-Master
**Branch with paper.md** (empty if default branch):
**Version:** v.1.0.0
