### 2021-12-02 eferrea/BCI-Master#2 @mstimberg
Make role of TrackM clearer
[Opened as part of JOSS review openjournals/joss-reviews#3956]

From the README, I wasn't entirely sure whether the TrackM is supposed to be basically used as it is, or whether it is just presented as an example for the real code that users  would write. If it is meant to be reused, then I find it quite inconvenient to change the source code just to change IP addresses (in particular, since the one that is given by default is an internal network address instead of something more generic like 127.0.0.1) – wouldn't it make sense to hand them over via command line arguments or a configuration file, for example? If this file is just an example, then I think more documentation is needed to describe what users would have to be aware of when writing such a program (link with VRPN, etc.).
