### 2021-12-02 eferrea/BCI-Master#3 @mstimberg
Improve documentation
[Opened as part of JOSS review openjournals/joss-reviews#3956]

The README file is currently the only existing documentation, and is a bit sparse (even though maybe a lot of this is very conventional for BCI software). In particular:
 * The elements of the GUI could need more information: Why are the units presented in 4 panels (only for space reasons?) and what are the axes (x = channel and y = unit?). How exactly is the "tuning strength" encoded in the colors (a colorbar might be helpful as well)? What do the columns in the table show?
 * The "extended functionalities" also merit some more explanation. In particular, it is sometimes unclear in what order these functionalities can be used: there is a "Load Decoder" but no "Save Decoder" – are decoders saved automatically? At what point should one change "Shared Control", etc. Also, I did not immediately understand the difference between "Check correlation" and "BCIIDLE".
 * The Matlab code for `BCI_Loop` has some comments how to call it (e.g. with suggestions for the update time and the number of neurons), but other functions/classes like `simNeurons_2D_velocity` do not state what their arguments are; this information is only included in individual comments in the source code itself.
 * Finally, the `BCImat` folder contains some basic tests in the `Test` folder, but it is unclear what these do (and they also have hardcoded IP addresses).
