### 2021-12-02 eferrea/BCI-Master#4 @mstimberg
 Automatic tests or test procedures
[Opened as part of JOSS review openjournals/joss-reviews#3956]

Obviously, automatic testing GUI software is complicated, so I don't expect the authors to write a test suite covering the full functionality of their software. It would be nice, however, to have clearer instructions on what is expected when using the artificial neurons. Something along the lines of "perform about 50 reaches in the task controller, then select about 20 units with a tuning strength above X. After pressing "Switch BCI", the cursor movement should go to the points on a fairly straight line; "Check correlation" should show correlations around Y". When I tested the software, it seemed to do something reasonable, but I wasn't really sure what to expect in detail.
