### 2021-12-02 eferrea/BCI-Master#6 @mstimberg
Community guidelines
[Opened as part of JOSS review openjournals/joss-reviews#3956]

The repository should mention where users should contact you, whether you are interested in contribution and what system (e.g. GitHub issues and Pull Requests) you use for bug reports and contributions.
