+++ README.md
@@ -14,7 +14,7 @@ BCImat is a Matlab GUI based program implementinhg a Brain-Computer Interface (B
 
 **Purpose of the task controller (TrackM)**
 
-TrackM is an example software written in C++ implementing a standard task controller for a reaching task to displayed targets. This part of the software can be replaced (or expanded) depending on which behavioral task one wants to be performed. Therefore, despite this not being the core of the project, users can test the full BCI closed-loop functionalities by running this project simultaneously with the BCI matlab framework. TrackM can also be written in any other programming language as far as the vrpn methods are used to send and read the data to and from the BCImat interface. The project contains a main.cpp together with a class implementation of the vrpn server method which should be build with VRPN (for streaming the data via network) and SFML libraries (for graphical displays of targets to be reached). 
+TrackM is an example software written in C++ implementing a standard task controller for a reaching task with the mouse pointer to displayed targets. This part of the software can be replaced (or expanded) depending on which behavioral task one wants to be performed. Therefore, despite this not being the core of the project, users can test the full BCI closed-loop functionalities by running the task controller simultaneously with the BCI matlab framework. This part of the project can also be written in any other programming language as far as the vrpn methods are used to send and read the data to and from the BCImat interface. TrackM contains a main.cpp together with a class implementation of the vrpn server method which should be build with VRPN (for streaming the data via network) and SFML libraries (for graphical displays of targets to be reached). 
 
 **Build the task controller project TrackM**
 
