+++ BCImat/2021-12-8_13-40-14_BCI_Task.txt
@@ -1,0 +1,1 @@
+  time        Trial                     Stage       BCI_on      IDLE_on        tgt_x        tgt_y        tgt_z          Hit       EYE_on
