+++ 2021-12-8_13-41-29_decoder.txt
@@ -1,355 +1,0 @@
-  time     K
-       0.002         0.000
-       0.061         0.000
-       0.113         0.000
-       0.168         0.000
-       0.219         0.000
-       0.270         0.000
-       0.321         0.000
-       0.370         0.000
-       0.420         0.000
-       0.471         0.000
-       0.522         0.000
-       0.572         0.000
-       0.623         0.000
-       0.675         0.000
-       0.726         0.000
-       0.777         0.000
-       0.827         0.000
-       0.884         0.000
-       0.934         0.000
-       0.986         0.000
-       1.042         0.000
-       1.092         0.000
-       1.142         0.000
-       1.192         0.000
-       1.242         0.000
-       1.292         0.000
-       1.342         0.000
-       1.393         0.000
-       1.444         0.000
-       1.495         0.000
-       1.545         0.000
-       1.595         0.000
-       1.646         0.000
-       1.697         0.000
-       1.747         0.000
-       1.797         0.000
-       1.847         0.000
-       1.897         0.000
-       1.948         0.000
-       1.999         0.000
-       2.050         0.000
-       2.101         0.000
-       2.151         0.000
-       2.201         0.000
-       2.251         0.000
-       2.301         0.000
-       2.351         0.000
-       2.401         0.000
-       2.451         0.000
-       2.501         0.000
-       2.552         0.000
-       2.602         0.000
-       2.652         0.000
-       2.703         0.000
-       2.754         0.000
-       2.804         0.000
-       2.854         0.000
-       2.904         0.000
-       2.954         0.000
-       3.005         0.000
-       3.055         0.000
-       3.106         0.000
-       3.157         0.000
-       3.215         0.000
-       3.266         0.000
-       3.316         0.000
-       3.367         0.000
-       3.417         0.000
-       3.467         0.000
-       3.518         0.000
-       3.568         0.000
-       3.618         0.000
-       3.668         0.000
-       3.718         0.000
-       3.769         0.000
-       3.819         0.000
-       3.869         0.000
-       3.919         0.000
-       3.983         0.000
-       4.033         0.000
-       4.085         0.000
-       4.135         0.000
-       4.185         0.000
-       4.236         0.000
-       4.288         0.000
-       4.339         0.000
-       4.389         0.000
-       4.439         0.000
-       4.490         0.000
-       4.539         0.000
-       4.590         0.000
-       4.641         0.000
-       4.692         0.000
-       4.742         0.000
-       4.793         0.000
-       4.844         0.000
-       4.895         0.000
-       4.946         0.000
-       4.997         0.000
-       5.047         0.000
-       5.098         0.000
-       5.148         0.000
-       5.198         0.000
-       5.249         0.000
-       5.299         0.000
-       5.349         0.000
-       5.399         0.000
-       5.450         0.000
-       5.500         0.000
-       5.660         0.000
-       5.710         0.000
-       5.760         0.000
-       5.811         0.000
-       5.861         0.000
-       5.912         0.000
-       5.963         0.000
-       6.013         0.000
-       6.064         0.000
-       6.115         0.000
-       6.166         0.000
-       6.216         0.000
-       6.266         0.000
-       6.317         0.000
-       6.368         0.000
-       6.418         0.000
-       6.468         0.000
-       6.519         0.000
-       6.570         0.000
-       6.622         0.000
-       6.672         0.000
-       6.722         0.000
-       6.772         0.000
-       6.824         0.000
-       6.873         0.000
-       6.924         0.000
-       6.975         0.000
-       7.025         0.000
-       7.075         0.000
-       7.126         0.000
-       7.177         0.000
-       7.228         0.000
-       7.278         0.000
-       7.329         0.000
[…]
