+++ 2021-12-8_13-40-34_decoder.txt
@@ -1,680 +1,0 @@
-  time     K
-       0.013         0.000
-       0.229         0.000
-       0.313         0.000
-       0.363         0.000
-       0.418         0.000
-       0.470         0.000
-       0.521         0.000
-       0.571         0.000
-       0.621         0.000
-       0.672         0.000
-       0.722         0.000
-       0.772         0.000
-       0.822         0.000
-       0.873         0.000
-       0.923         0.000
-       0.974         0.000
-       1.024         0.000
-       1.074         0.000
-       1.124         0.000
-       1.174         0.000
-       1.230         0.000
-       1.280         0.000
-       1.330         0.000
-       1.380         0.000
-       1.431         0.000
-       1.481         0.000
-       1.531         0.000
-       1.582         0.000
-       1.632         0.000
-       1.682         0.000
-       1.732         0.000
-       1.783         0.000
-       1.833         0.000
-       1.883         0.000
-       1.933         0.000
-       1.983         0.000
-       2.033         0.000
-       2.083         0.000
-       2.133         0.000
-       2.183         0.000
-       2.234         0.000
-       2.284         0.000
-       2.334         0.000
-       2.385         0.000
-       2.435         0.000
-       2.485         0.000
-       2.535         0.000
-       2.585         0.000
-       2.635         0.000
-       2.685         0.000
-       2.735         0.000
-       2.785         0.000
-       2.835         0.000
-       2.885         0.000
-       2.935         0.000
-       2.985         0.000
-       3.035         0.000
-       3.086         0.000
-       3.135         0.000
-       3.186         0.000
-       3.236         0.000
-       3.286         0.000
-       3.337         0.000
-       3.387         0.000
-       3.437         0.000
-       3.488         0.000
-       3.538         0.000
-       3.589         0.000
-       3.640         0.000
-       3.691         0.000
-       3.741         0.000
-       3.792         0.000
-       3.843         0.000
-       3.893         0.000
-       3.944         0.000
-       3.995         0.000
-       4.045         0.000
-       4.096         0.000
-       4.147         0.000
-       4.197         0.000
-       4.248         0.000
-       4.298         0.000
-       4.348         0.000
-       4.398         0.000
-       4.448         0.000
-       4.498         0.000
-       4.548         0.000
-       4.599         0.000
-       4.649         0.000
-       4.700         0.000
-       4.750         0.000
-       4.801         0.000
-       4.851         0.000
-       4.902         0.000
-       4.953         0.000
-       5.003         0.000
-       5.053         0.000
-       5.103         0.000
-       5.153         0.000
-       5.203         0.000
-       5.253         0.000
-       5.303         0.000
-       5.354         0.000
-       5.418         0.000
-       5.469         0.000
-       5.521         0.000
-       5.572         0.000
-       5.629         0.000
-       5.686         0.000
-       5.736         0.000
-       5.787         0.000
-       5.838         0.000
-       5.888         0.000
-       5.937         0.000
-       5.988         0.000
-       6.037         0.000
-       6.093         0.000
-       6.143         0.000
-       6.194         0.000
-       6.244         0.000
-       6.294         0.000
-       6.344         0.000
-       6.395         0.000
-       6.446         0.000
-       6.497         0.000
-       6.548         0.000
-       6.600         0.000
-       6.665         0.000
-       6.715         0.000
-       6.765         0.000
-       6.815         0.000
-       6.865         0.000
-       6.916         0.000
-       6.966         0.000
-       7.016         0.000
-       7.066         0.000
-       7.116         0.000
-       7.167         0.000
-       7.217         0.000
-       7.267         0.000
-       7.317         0.000
-       7.367         0.000
-       7.418         0.000
-       7.468         0.000
[…]
