### 2021-12-08 eferrea/BCI-Master#2 @eferrea
Hi @mstimberg, I updated the text in the readme file to clarify the usage of the TrackM. In addition as you suggested, I add a configuration file where to specify the information relative to addresses of the tracked objects.

Here what was added in the text.

Purpose of the task controller (TrackM)

TrackM is an example software written in C++ implementing a standard task controller for a reaching task with the mouse pointer to displayed targets. This part of the software can be replaced (or expanded) depending on which behavioral task one wants to be performed. Therefore, despite this not being the core of the project, users can test the full BCI closed-loop functionalities by running the task controller simultaneously with the BCI matlab framework. This part of the project can also be written in any other programming language as far as the vrpn methods are used to send and read the data to and from the BCImat interface. TrackM contains a main.cpp together with a class implementation of the vrpn server method which should be build with VRPN (for streaming the data via network) and SFML libraries (for graphical displays of targets to be reached).

Running the BCI framework

Edit the provided configuration file by entering a name for the Tracker (mouse pointer in this case) implemented in the task controller (trackM) followed by @<server_address>. Also a name for the Tracker (output of the BCI decoder) implemented in the BCI server is needed and followed by @<client_address>. Here TrackerTC and TrackerBCI are used as names followed by their IP adresses. They specify the adresses of the computers were the task controller and the BCImat framework run.
Note that the task controller streams the information of the mouse pointer (TrackerTC) by implementing a vrpn server to the BCI framework and read the information provided by the BCI framework (via TrackerBCI) by implementing a vrpn client. Also a name of a port needs to be added to avoid conflictts when task controller and BCI framework run on the same computer. The provided entry <6666> can be left unchanged. Also the number of dots per inch for the specific used screen should be specified for pixel to mm conversions (not critical) In summary the configuration file should look like that:
IP_address_server = TrackerTC@127.0.0.1
IP_address_client = TrackerBCI@127.0.0.1
port = 6666
dpi = 108.79

Run the task controller program TrackM

Run the function BCI_loop.m inside BCImat with the following arguments:

for the simulation mode: BCI_Loop(isBrain,neurons,BCI_update_time,delay,server_address,client_address,port)
practically it will look like that:

BCI_Loop(false,60,0.05,0,'TrackerBCI@172.17.6.10','TrackerTC@172.17.6.10',6666)

I will go through the other questions in the next days.

Regards,
Enrico
