+++ README.md
@@ -83,7 +83,7 @@ Additional functionalities are listed in the next paragraph.
 
 ![BCImat](https://user-images.githubusercontent.com/40661882/125582844-48d7406e-c0f1-404a-8047-a63615ed8ab2.png)
 
-*Fig 1.	The GUI layout reflects our experimental settings.  We used six electrode arrays each of them containing 32 channels (total 128 of recording sites). The online recorded spikes were streamed with 128 channel Cerebus recording system (Blackrock Microsystems, Salt Lake City, USA, https://www.blackrockmicro.com/). In the GUI, the identity of the channel is arranged as column while each row represent a different unit for that channel. The tested  recording system streams up to six units per channel that can be identified online. The table in the middle show the tuning properties of the selected unit by maintaining the same numenclature as explained previously (colum = channel identity, row = unit identity)* .
+*Fig 1.	The adopted GUI layout reflects our experimental settings.  We used six electrode arrays each of them containing 32 channels (total 128 of recording sites). The online recorded spikes were streamed with 128 channel Cerebus recording system (Blackrock Microsystems, Salt Lake City, USA, https://www.blackrockmicro.com/). In the top left part of the GUI, the identity of the channel is arranged as column while each row represent a different unit for that channel. We used six rows given that the tested  recording system streams a maximum of up to six units per channel that can be online identified.  Here the numenclature is maintained as explained previously (ch = channel identity, U = unit identity)* . The table in the middle shows the tuning properties at specic calibration interval of the decoder.
 
 
 **Matlab Graphical user interface extended functionalities**
