+++ README.md
@@ -83,7 +83,7 @@ Additional functionalities are listed in the next paragraph.
 
 ![BCImat](https://user-images.githubusercontent.com/40661882/125582844-48d7406e-c0f1-404a-8047-a63615ed8ab2.png)
 
-*Fig 1.	The adopted GUI layout reflects our experimental settings.  We used six electrode arrays each of them containing 32 channels (total 128 of recording sites). The online recorded spikes were streamed with 128 channel Cerebus recording system (Blackrock Microsystems, Salt Lake City, USA, https://www.blackrockmicro.com/). In the top left part of the GUI, the identity of the channel is arranged as column while each row represent a different unit (or neuron) for that channel. We used six rows given that the tested  recording system streams a maximum of six units per channel which can be online identified. In the middle column the name of the selectcted unit is displayed (ch = channel identity, U = unit identity)* . The column in the middle shows the tuning properties at specic calibration interval of the decoder (samples/Hit).
+*Fig 1.	Graphical user interface exploiting BCI functionalities. The GUI layout reflects our experimental settings.  We used six electrode arrays each of them containing 32 channels (total 128 of recording sites). The online recorded spikes were streamed with 128 channel Cerebus recording system (Blackrock Microsystems, Salt Lake City, USA, https://www.blackrockmicro.com/). In the top left part of the GUI, the identity of the channel is arranged as column while each row represent a different unit (or neuron) for that channel. We used six rows given that the tested  recording system streams a maximum of six units per channel which can be online identified. In the middle column the name of the selectcted unit is displayed (ch = channel identity, U = unit identity)* . The column in the middle shows the tuning properties calculated after the specific calibration intervals (samples/Hit) of the decoder.
 
 
 **Matlab Graphical user interface extended functionalities**
