+++ README.md
@@ -83,7 +83,7 @@ Additional functionalities are listed in the next paragraph.
 
 ![BCImat](https://user-images.githubusercontent.com/40661882/125582844-48d7406e-c0f1-404a-8047-a63615ed8ab2.png)
 
-*Fig 1.	Graphical user interface exploiting BCI functionalities. The GUI layout on the left displays all possible recorded units arranged to reflect our experimental settings. Since we used six electrode arrays each of them containing 32 channels (total 128 of recording sites) we decided to clearly separate each array from the other in the visualization. The recorded spikes were online streamed with a 128 channel Cerebus recording system (Blackrock Microsystems, Salt Lake City, USA, https://www.blackrockmicro.com/). The identity of the channel is arranged as column while each row represent a different unit (or spike) for that channel. For each electrode array, we split the data in six rows given that the tested  recording system streams a maximum of six units per channel which can be online identified. In the middle column the name of the selectcted unit is displayed (ch = channel identity, U = unit identity)* . The column in the middle shows the tuning properties calculated after the specific calibration intervals (samples/Hit) of the decoder.
+*Fig 1.	Graphical user interface exploiting BCI functionalities. The GUI layout on the left displays all possible recorded units arranged to reflect our experimental settings. Since we recorded simultaneously from four electrode arrays each of them containing 32 channels (total 128 of recording sites), we clearly separate each array from the other in the visualization. The identity of the channel is arranged as column while each row represents a different unit (or spike) for that channel. For each electrode array, we split the data in six rows given that the recording system streams a maximum of six units (after online sorting). In the middle column the name of the selectcted unit is displayed (ch = channel identity, U = unit identity)* together with the its tuning properties calculated after the specific calibration intervals (samples/Hit) of the decoder. R represents the explained variance (R2) of fitting a cosine tuned model to the firing rate of each single cell. This value is also color coding the GUI on the left column. For this reason strongly tuned cells are identified with warmer columns.  
 
 
 **Matlab Graphical user interface extended functionalities**
