### 2021-12-14 760d885
Review Joss 1_3
Changed the order of buttons according to practical usage
