+++ BCImat/BCI_Classes/simNeurons_2D_velocity.m
@@ -3,6 +3,8 @@
 %baseline rate of a specified number of cells. The cell firing dynamic is
 %dictate by a cosine model (Georgopoulos et al., 1982).
 % @ E.Ferrea ,2015 .
+%At the actual implementation status, this class takes only the number of units as
+%input. 
 
 classdef simNeurons_2D_velocity < handle  % class should be implemented as a handle class. 
     % All copies of a given handle object refer to the same data.
