+++ BCImat/2021-12-8_13-40-14_BCI_trackers.txt
@@ -1,1 +1,0 @@
-  time        pos_x        pos_y        vel_x        vel_y    BCI_pos_x    BCI_pos_y    BCI_vel_x    BCI_vel_y
