+++ README.md
@@ -122,7 +122,7 @@ the number of units should correspond as well as their tuning characteristics. U
  * Kalman_decoder_class.m: it uses the calibration matrix generated by the the Kalman_calibration class to generate position estimation online. Ideally this class and the calibrator class could be replaced with any other decoder type implementing a similar behaviour. 
 
 2. Inside  Test_connections:
- * TestVRPNConnection.m: is used to simply test if a client VRPN connection can be established without implementing all the BCI part
+ * TestVRPNConnection.m: is used to test if a VRPN client connection can be established without implementing all the BCI closed-loop.
  * Test_Blackrock: is used to test if a connection can be established with the Cereplex system.
 
   
