### 2021-12-28 eferrea/BCI-Master#3 @eferrea
Hi @mstimberg,
* In the figure legend I better explain, how users should interpret the GUI features and the reason why the units were separated in four matrixes. Unfortunately, when I tested the color bar it was slowing down the update of the online display. The value of the color can be still read from the R2 column.
* An better explanation of the extended functionalities was added in the readme file.
* Other contents of the package like sumNeurons_2D_velocity were also better explained in a dedicated paragraph in the readme file.
* They are used for simple tests of VRPN connection and  the cbmex.m connection (application mode). I renamed the folder Test_connections for clarity. An explanation of the content of this folder is also provided in the readme now.  The code has also been cleaned and generic local host IP addresses used.
