### 2021-12-28 eferrea/BCI-Master#4 @eferrea
Hi @mstimberg,
a dedicated  "test procedure (simulation mode)" paragraph was added to the readme file with a detailed explanation of what the user should do and what should expect during the first  session.
