### 2022-01-03 fc99c46
makelist added
Added makelist instead of visual Studio project. Tested linux and Windows
