+++ TrackM/build/CMakeFiles/3.22.1/CMakeCXXCompiler.cmake
@@ -1,83 +1,0 @@
-set(CMAKE_CXX_COMPILER "C:/Program Files (x86)/Microsoft Visual Studio 12.0/VC/bin/x86_amd64/cl.exe")
-set(CMAKE_CXX_COMPILER_ARG1 "")
-set(CMAKE_CXX_COMPILER_ID "MSVC")
-set(CMAKE_CXX_COMPILER_VERSION "18.0.40629.0")
-set(CMAKE_CXX_COMPILER_VERSION_INTERNAL "")
-set(CMAKE_CXX_COMPILER_WRAPPER "")
-set(CMAKE_CXX_STANDARD_COMPUTED_DEFAULT "98")
-set(CMAKE_CXX_EXTENSIONS_COMPUTED_DEFAULT "OFF")
-set(CMAKE_CXX_COMPILE_FEATURES "cxx_std_98;cxx_std_11;cxx_std_14;cxx_std_17;cxx_std_20;cxx_std_23;cxx_template_template_parameters;cxx_alias_templates;cxx_auto_type;cxx_decltype;cxx_default_function_template_args;cxx_defaulted_functions;cxx_delegating_constructors;cxx_enum_forward_declarations;cxx_explicit_conversions;cxx_extended_friend_declarations;cxx_extern_templates;cxx_final;cxx_generalized_initializers;cxx_lambdas;cxx_local_type_template_args;cxx_long_long_type;cxx_nullptr;cxx_override;cxx_range_for;cxx_raw_string_literals;cxx_right_angle_brackets;cxx_rvalue_references;cxx_static_assert;cxx_strong_enums;cxx_trailing_return_types;cxx_uniform_initialization;cxx_variadic_macros;cxx_variadic_templates;cxx_contextual_conversions")
-set(CMAKE_CXX98_COMPILE_FEATURES "")
-set(CMAKE_CXX11_COMPILE_FEATURES "")
-set(CMAKE_CXX14_COMPILE_FEATURES "")
-set(CMAKE_CXX17_COMPILE_FEATURES "")
-set(CMAKE_CXX20_COMPILE_FEATURES "")
-set(CMAKE_CXX23_COMPILE_FEATURES "")
-
-set(CMAKE_CXX_PLATFORM_ID "Windows")
-set(CMAKE_CXX_SIMULATE_ID "")
-set(CMAKE_CXX_COMPILER_FRONTEND_VARIANT "")
-set(CMAKE_CXX_SIMULATE_VERSION "")
-set(CMAKE_CXX_COMPILER_ARCHITECTURE_ID x64)
-
-set(MSVC_CXX_ARCHITECTURE_ID x64)
-
-set(CMAKE_AR "C:/Program Files (x86)/Microsoft Visual Studio 12.0/VC/bin/x86_amd64/lib.exe")
-set(CMAKE_CXX_COMPILER_AR "")
-set(CMAKE_RANLIB ":")
-set(CMAKE_CXX_COMPILER_RANLIB "")
-set(CMAKE_LINKER "C:/Program Files (x86)/Microsoft Visual Studio 12.0/VC/bin/x86_amd64/link.exe")
-set(CMAKE_MT "CMAKE_MT-NOTFOUND")
-set(CMAKE_COMPILER_IS_GNUCXX )
-set(CMAKE_CXX_COMPILER_LOADED 1)
[…]
