+++ README.md
@@ -14,13 +14,23 @@ BCImat is a Matlab GUI based program implementinhg a Brain-Computer Interface (B
 
 **Purpose of the task controller (TrackM)**
 
-TrackM is an example software written in C++ implementing a standard task controller for a reaching task with the mouse pointer to displayed targets. This part of the software can be replaced (or expanded) depending on which behavioral task one wants to be performed. Therefore, despite this not being the core of the project, users can test the full BCI closed-loop functionalities by running the task controller simultaneously with the BCI matlab framework. This part of the project can also be written in any other programming language as far as the vrpn methods are used to send and read the data to and from the BCImat interface. TrackM contains a main.cpp together with a class implementation of the vrpn server method which should be build with VRPN (for streaming the data via network) and SFML libraries (for graphical displays of targets to be reached). 
+TrackM is an example software written in C++ implementing a standard task controller for a reaching task with the mouse pointer to displayed targets. The task controller allows users to perform sequential reacheas to  a target (green circle) starting from a central fixation circle (gray) with the mouse. This part of the software can be replaced (or expanded) depending on which behavioral task one wants to be performed. Therefore, despite this not being the core of the project, users can test the full BCI closed-loop functionalities by running the task controller simultaneously with the BCI matlab framework. This part of the project can also be written in any other programming language as far as the vrpn methods are used to send and read the data to and from the BCImat interface. TrackM contains a main.cpp together with a class implementation of the vrpn server method (vrpn_server_class) which should be build with VRPN (for streaming the data via network) and SFML libraries (for graphical displays of targets to be reached). 
 
 **Build the task controller project TrackM**
 
-BCImat comes with a simple task controller written in c++ to interface with the BCImat. The task controller allows users to perform sequential reacheas to  a target (green circle) starting from a central fixation circle (gray).
-The c++ project requires the graphic library SFML (available at https://www.sfml-dev.org/download.php) and the virtual reality peripheral network library (VRPN) (available at https://github.com/vrpn/vrpn/wiki) to be linked to the project. 
-The project contains a main.cpp running the task including the vrpn client callback functions (similar to http://www.vrgeeks.org/vrpn/tutorial---use-vrpn) as well as an implemented vrpn server class. These files are contained in the "Source Files" folder of the project and can be also used to make a project from scratch in other integrated development environment (we also tested Xcode). 
+The  task controller folder TrackM contains the source code and a CMakeLists.txt to build the project under different OS architectures. 
+Prior to use cmake to generate the build environment, TrackM requires the graphic library SFML (available at https://www.sfml-dev.org/download.php) and the virtual reality peripheral network library (VRPN) (available at https://github.com/vrpn/vrpn) to be installed on your computer. 
+After doing that, in Windows it is neccesary to specify the path of the include  and library folders of SFML and VRPN libraries. Therefore, in the CMakeLists.txt you have to change the content of the SET command to match the full path of your include and lib directories for SFML and VRPN (5 lines in total).
+In practice what you have to change in the CMakeLists.txt is the following:
+SET(SFML_INCLUDE_PATH <change to the full path of your SFML include directory>)
+SET(VRPN_INCLUDE_PATH <change to the full path of your VRPN include directory>)
+SET(SFML_LIBRARY_PATH <change to the full path of your SFML lib directory>)
+SET(VRPN_LIBRARY_PATH <change here to the full path containing vrpn.lib>)
+SET(VRPN_QUAT_LIBRARY_PATH <change here to the full path containing quat.lib>)
+ 
+In Linux, it is not neccessary to specify these folders unless libraries are installed in custom locations.
+
[…]
