+++ README.md
@@ -21,9 +21,9 @@ TrackM is an example software written in C++ implementing a standard task contro
 The  task controller folder TrackM contains the source code and a CMakeLists.txt to build the project under different OS architectures. 
 Prior to use cmake to generate the build environment, TrackM requires the graphic library SFML (available at https://www.sfml-dev.org/download.php) and the virtual reality peripheral network library (VRPN) (available at https://github.com/vrpn/vrpn) to be installed on your computer. 
 After doing that, in Windows it is neccesary to specify the path of the include  and library folders of SFML and VRPN libraries. Therefore, in the CMakeLists.txt you have to change the content of the SET command to match the full path of your include and lib directories for SFML and VRPN (5 lines in total).
-In practice what you have to change in the CMakeLists.txt is the following:
-SET(SFML_INCLUDE_PATH <change to the full path of your SFML include directory>)  
- SET(VRPN_INCLUDE_PATH <change to the full path of your VRPN include directory>)  
+In practice what you have to change in the CMakeLists.txt is the following:  
+SET(SFML_INCLUDE_PATH <change to the full path of your SFML include directory>)   
+SET(VRPN_INCLUDE_PATH <change to the full path of your VRPN include directory>)  
 SET(SFML_LIBRARY_PATH <change to the full path of your SFML lib directory>)  
 SET(VRPN_LIBRARY_PATH <change here to the full path containing vrpn.lib>)  
 SET(VRPN_QUAT_LIBRARY_PATH <change here to the full path containing quat.lib>)
