+++ README.md
@@ -32,6 +32,8 @@ In Linux, it is not neccessary to specify these folders unless libraries are ins
 
 In Windows, after making and bulding the Project, don't forget to copy the SFML DLLs (they are in <sfml-install-path/bin>) into the folder containing your executable (see also https://www.sfml-dev.org/tutorials/2.5/start-vc.php) 
 
+At the end both in Linux and Windows, copy the config.txt contained in the TrackM folder into into the folder containing your executable.
+
 
 **VRPN matlab client and server for BCImat**
 
