+++ CONTRIBUTING.md
@@ -1,0 +1,6 @@
+**How to contribute**
+
+**Questions:** please create a new issue together with your question.  
+**Bugs:** known bugs are left as open issues on GitHub, if you find a new one, please report it there with the flag bugs.  
+**Enhancements:** please create a new issue with the flag enhancements.  
+**Modifications:** You are very welcome to contribute to this project! For discussion create a new issue and if your modifications are ready just submit a pull request. You should make sure that your changes do not break anything by running BCI_loop.m in combination with TrackM.  I will review the code and its functionality and eventually discuss/make changes to your commit before merging when it is working satisfactorily.
