+++ README.md
@@ -12,7 +12,7 @@ BCImat is a Matlab GUI-based program implementinhg a Brain-Computer Interface (B
 1. a simulated set of cosine tuned neurons for testing purposes(simulation mode) before using it under real brain control,
 2. a real neural interface using Blackrock 128 channel recording system (Blackrock Microsystems, Salt Lake City, USA, https://www.blackrockmicro.com/) via their cbmex code for real intracortical control (application mode).  
 
-In the actual version, the "Statistics and Machine Learning Toolbox" is needed. In future releases, I will remove functionalities from this toolbox.
+In the actual version, the "Statistics and Machine Learning Toolbox" is needed. In future releases, I will remove dependencies from this toolbox.
 
 **Purpose of the task controller (TrackM)**
 
