+++ README.md
@@ -4,7 +4,7 @@
 **Content of the package**
 
 1. A Matlab BCI framework (BCImat).
-2. A c++ project implementing a task controller for testing the Matlab BCI framework  (TrackM).
+2. A c++ project (TrackM) implementing a task controller for testing the Matlab BCI framework in closed-loop fashion.
 
 **Purpose of the BCI framework (BCIMat)**
 
@@ -16,13 +16,13 @@ In the actual version, the "Statistics and Machine Learning Toolbox" is needed. 
 
 **Purpose of the task controller (TrackM)**
 
-TrackM is an example software written in C++ implementing a standard task controller for a reaching task with the mouse pointer to displayed targets. The task controller allows users to perform sequential reacheas to  a target (green circle) starting from a central fixation circle (gray) with the mouse. This part of the software can be replaced (or expanded) depending on which behavioral task one wants to be performed. Therefore, despite this not being the core of the project, users can test the full BCI closed-loop functionalities by running the task controller simultaneously with the BCI matlab framework. This part of the project can also be written in any other programming language as far as the vrpn methods are used to send and read the data to and from the BCImat interface. TrackM contains a main.cpp together with a class implementation of the vrpn server method (vrpn_server_class) which should be build with VRPN (for streaming the data via network) and SFML libraries (for graphical displays of targets to be reached). 
+TrackM is an example software written in C++ implementing a standard task controller for a reaching task. The task controller allows users to perform sequential reaches to  a target (green circle) starting from a central fixation circle (gray) with the mouse. This part of the software can be replaced (or expanded) depending on which behavioral task one wants to be performed. Therefore, despite this not being the core of the project, users can test the full BCI closed-loop functionalities by running the task controller simultaneously with the BCI Matlab framework. This part of the project can also be written in any other programming language as far as the VRPN methods are used to send and read the data to and from the BCImat interface. TrackM contains a main.cpp together with a class implementation of the vrpn server method (vrpn_server_class) which should be built with VRPN (for streaming the data via network) and SFML libraries (for graphical displays of targets to be reached). 
 
 **Build the task controller project TrackM**
 
-The  task controller folder TrackM contains the source code and a CMakeLists.txt to build the project under different OS architectures. 
-Prior to use cmake to generate the build environment, TrackM requires the graphic library SFML (available at https://www.sfml-dev.org/download.php) and the virtual reality peripheral network library (VRPN) (available at https://github.com/vrpn/vrpn) to be installed on your computer. 
-After doing that, in Windows it is neccesary to specify the path of the include  and library folders of SFML and VRPN libraries. Therefore, in the CMakeLists.txt you have to change the content of the SET command to match the full path of your include and lib directories for SFML and VRPN (5 lines in total).
+The task controller folder TrackM contains the source code and a CMakeLists.txt to build the project under different OS architectures. 
+Before using cmake to generate the build environment, TrackM requires the graphic library SFML (available at https://www.sfml-dev.org/download.php) and the virtual reality peripheral network library (VRPN) (available at https://github.com/vrpn/vrpn) to be installed on your computer. 
+After doing that, in Windows, it is necessary to specify the path of the include and library folders of SFML and VRPN libraries. Therefore, in the CMakeLists.txt you have to change the content of the SET command to match the full path of your include and lib directories for SFML and VRPN (5 lines in total).
 In practice what you have to change in the CMakeLists.txt is the following:  
[…]
