+++ README.md
@@ -117,23 +117,23 @@ These GUIs include:
 4) *Check Correlation*: to be used in BCIIDLE mode, by pressing this button the Pearson correlation coefficient is estimated among decoded speeds and real speeds. This button is mainly intended for the application mode where before to allowing users to control the decoder, a decent (offline) performance needs to be achieved.
 5) *Switch BCI*: after successful calibration to switch to close loop control. It switches to the BCI closed-loop mode. A decent amount of units (30-40) and a decent amount of samples (200-300) needs to be acquired to achieve a decent performance (in offline mode the green cursor should be reliably controlled with the mouse pointer).
 6) *Update Decoder*: to be used after a recalibration in closed-loop mode with Updated Regression. Important for real applications not in the simulation mode.  
-7) *Load Decoder*: to load a previous calibrated decoder. Note that if used among different experimental sessions
-the number of units should correspond as well as their tuning characteristics. Useful for now to restore previous decoders in the same session but potentially useful to use the same decoder across many days.
-8) *Stop BCI*: stop the program loop. Should this operation not being successfull, to restart the BCImat the vrpn server needs to be manually stopped by doing the following in the Matlab shell:  *vrpn_server('stop_server')*
+7) *Load Decoder*: to load a previously calibrated decoder. Note that if used among different experimental sessions
+the same number of units should be maintained. Useful for now to restore previous decoders in the same session but potentially useful to use the same decoder across many days.
+8) *Stop BCI*: stop the program loop. Should this operation not being successful, to restart the BCImat the VRPN server needs to be manually stopped by doing the following in the Matlab shell:  *vrpn_server('stop_server')*
 9) *Reset Calibrator*: to reset the calibration if something went wrong during calibration. It starts to recollect speed and neural samples.  
-10) *Shared control*: the cursor control is shared among the computer directly pointing at the target and the neurons. Specify a number between 0-1. 1: full computer control. It is useful during subject' training phases and to recalibrate the decoder during closed loop control. The idea is that during real experiments a high level of computer control is introduced to maintain high performance. During the task as subject acquire proficiency with the control the amount of control from the computer side is reduced.
-11) *Perturbation panel*: rotate units preferred direction . It needs to specify the angle of rotation as well as the percentage of random units that will be rotated and start the perturbation. By rotating the preferred direction of some units during closed-loop control, the direction of movement deviates from the intended movement directions. This panel is intended to introduce visuo-motor rotations of the cursors useful for motor learning studies.
+10) *Shared control*: the cursor control is shared among the computer directly pointing at the target and the neurons. Specify a number between 0-1. 1: full computer control. It is useful during the subject' training phases and to recalibrate the decoder during closed-loop control. The idea is that during real experiments a high level of computer control is introduced to maintain high performance. During the task as the subject acquires proficiency with the control, the amount of control from the computer side is reduced.
+11) *Perturbation panel*: rotate the preferred direction of a subset of units. It needs to specify the angle of rotation as well as the percentage of random units that will be rotated and start the perturbation. By rotating the preferred direction of some units during closed-loop control, the direction of movement deviates from the intended movement directions. This panel is intended to introduce visuo-motor rotations of the cursors useful for motor learning studies.
 
 **Other contents of the package**
 
 1. Inside BCI_classes folder: 
 
[…]
