+++ README.md
@@ -22,7 +22,7 @@ TrackM is an example software written in C++ implementing a standard task contro
 
 The task controller folder TrackM contains the source code and a CMakeLists.txt to build the project under different OS architectures. 
 Before using cmake to generate the build environment, TrackM requires the graphic library SFML (available at https://www.sfml-dev.org/download.php) and the virtual reality peripheral network library (VRPN) (available at https://github.com/vrpn/vrpn) to be installed on your computer. 
-After doing that, in Windows, it is necessary to specify the path of the include and library folders of SFML and VRPN libraries. Therefore, in the CMakeLists.txt you have to change the content of the SET command to match the full path of your include and lib directories for SFML and VRPN (5 lines in total).
+After installing them, you can use cmake to generate the build environement. In Windows, it is necessary to specify  in the CmakeLists.txt the path of the include and library folders of SFML and VRPN libraries. Therefore, in the CMakeLists.txt you have to change the content of the SET command to match the full path of your include and lib directories for SFML and VRPN (5 lines in total).
 In practice what you have to change in the CMakeLists.txt is the following:  
 SET(SFML_INCLUDE_PATH *\<change to the full path of your SFML include directory\>*)   
 SET(VRPN_INCLUDE_PATH *\<change to the full path of your VRPN include directory\>*)  
@@ -32,7 +32,7 @@ SET(VRPN_QUAT_LIBRARY_PATH *\<change to the full path containing quat.lib\>*)
  
 In Linux, it is not necessary to specify these folders unless libraries are installed in custom locations.
 
-In Windows, after making and building the Project, don't forget to copy the SFML DLLs (they are in <sfml-install-path/bin>) into the folder containing your executable (see also https://www.sfml-dev.org/tutorials/2.5/start-vc.php) 
+In Windows, after "making" and "building" the Project, don't forget to copy the SFML DLLs (they are in <sfml-install-path/bin>) into the folder containing your executable (see also https://www.sfml-dev.org/tutorials/2.5/start-vc.php) 
 
 At the end, both in Linux and Windows, copy the config.txt contained in the TrackM folder into the folder containing your executable.
 
