+++ README.md
@@ -128,7 +128,7 @@ the same number of units should be maintained. Useful for now to restore previou
 
 1. Inside BCI_classes folder: 
 
- * simNeurons_2D_velocity.m: this class is used inside BCI_loop.m to generate firing rates of artificial neurons according to a Poisson distribution. During each small movement (e.g 50 ms timestamp), the mean of each Poisson neuron is determined by the baseline firing rate plus the cosine of the actual moving direction and the neuron preferred direction multiplied by the modulation depth. Each neuron preferred direction is determined according to a circular uniform distribution.  This class also uses a hard coded modulation depth in the physiological range of 4-18 Hz and a baseline firing rate in the range 4-20.  Future realeases of this class could include the possibility of specifying these ranges. ###
+ * simNeurons_2D_velocity.m: this class is used inside BCI_loop.m to generate firing rates of artificial neurons according to a Poisson distribution. During each small movement (e.g 50 ms timestamp), the mean of each Poisson neuron is determined by the baseline firing rate plus the cosine of the actual moving direction and the neuron preferred direction multiplied by the modulation depth. Each neuron preferred direction is determined according to a circular uniform distribution.  This class also uses a hard coded modulation depth in the physiological range of 4-18 Hz and a baseline firing rate in the range 4-20.  Future realeases of this class could include the possibility of specifying these ranges. 
 
  * task_parser.m: this simple class is used to handle messages coming from the task controller. It is important for the BCImat to know information about the state of the task    to adapt its behavior accordingly. For example, it could be useful to collect samples of firing rates and positional samples only when the users are engaged with the motor task. In this case, we decided to store the values for calibration online during the last stage of the task (stage 3). 
 
