### 2022-01-04 eferrea/BCI-Master#5 @eferrea
Hi @mstimberg, I provide now only the source code and a cmake list to build the project under different OS architectures. I tested the project under Windows and Ubuntu lts 20.04. The readme was also modified accordingly on how to use cmake in the different OSs.
I also added that the Matlab code requires the Statistics and Machine Learning toolbox. Given the few dependencies from this toolbox, I plan to remove them soon.
Finally, in the readme, I also specify to add the subfolders to the Matlab path.
I hope this can help speeding-up installation procedures.
