### 2022-01-04 eferrea/BCI-Master#6 @eferrea
Hi @mstimberg, thanks for pointing this out. A CONTRIBUTING.md file was added to the repository with instructions on how to contribute to this project.
