+++ paper/paper.bib
@@ -47,6 +47,7 @@
    author = {Ferrea, Enrico and Morel, Pierre and Franke, Joscha and Gail, Alexander},
    title = {Statistical determinants of visuomotor adaptation in a virtual reality three-dimensional environment},
    journal = {bioRxiv},
+   DOI = {10.1101/2021.02.16.431440},
    year = {2021},
    type = {Journal Article}
 }
@@ -59,6 +60,7 @@
    number = {11},
    pages = {1527-1537},
    ISSN = {0270-6474},
+   DOI = {doi.org/10.1523/JNEUROSCI.02-11-01527.1982},
    year = {1982},
    type = {Journal Article}
 }
@@ -161,6 +163,7 @@
    number = {2},
    pages = {101-115},
    ISSN = {2168-2291},
+   DOI = {10.1109/THMS.2020.2968411},
    year = {2020},
    type = {Journal Article}
 }
@@ -169,7 +172,9 @@
    author = {Taylor, Russell M and Hudson, Thomas C and Seeger, Adam and Weber, Hans and Juliano, Jeffrey and Helser, Aron T},
    title = {VRPN: a device-independent, network-transparent VR peripheral system},
    booktitle = {Proceedings of the ACM symposium on Virtual reality software and technology},
+   DOI = {10.1145/505008.505019},
    pages = {55-61},
+   year = {2001},
    type = {Conference Proceedings}
 }
 
@@ -181,6 +186,7 @@
    number = {1},
    pages = {80-118},
    ISSN = {0899-7667},
+   DOI = {10.1162/089976606774841585},
    year = {2006},
    type = {Journal Article}
 }
