### 2022-01-11 eferrea/BCI-Master#5 @mstimberg
Thanks for all these changes. I was able to build everything on Ubuntu Linux with cmake and without further manual configuration.
