### 2022-01-11 eferrea/BCI-Master#4 @mstimberg
Thanks, this is much easier to follow now. In Point 4 of the test procedure you say:
> After collecting enough samples (samples are shown on the right table, 150-200 for good performance) press the Switch BCI button to start the decoder.

Could you maybe add that "collecting samples" here means to continue to perform the reaching task, and that the samples are only updated when you press "Update Regression" again? Of course, only if that is actually the case :)
