### 2022-01-11 eferrea/BCI-Master#3 @mstimberg
Great, that is indeed much clearer. I cannot judge all the details, but at least everything seems to be mentioned.
Minor nitpick: The README states: "...extensively explained in the *Matlab Graphical user interface extended functionalities* paragraph ..." but the paragraph heading is called "Matlab Graphical user interface functionalities" (i.e. without "extended")
