### 2022-01-11 eferrea/BCI-Master#2 @mstimberg
Thanks for the additional explanations, this is much clearer now. The config file makes everything much more convenient as well.

If I am not mistaken, there is a minor typo here:
> and with the following arguments in case of application mode: BCI_Loop(true,60,0.05,0,'TrackerBCI@127.0.0.1','TrackerBCI@127.0.0.1',6666)

The second address argument should be "TrackerTC" instead of "TrackerBCI", right?
