### 2022-01-11 eferrea/BCI-Master#3 @mstimberg
One more very minor thing: it would be good to include the full reference for "Wu et al. 2006" in the README file (and not only in the paper) to make it self-contained.
