### 2022-01-11 eferrea/BCI-Master#2 @eferrea
Thanks for your question.
Yes, you are right. It should be:
BCI_Loop(false,60,0.05,0,'TrackerBCI@172.17.6.10','TrackerTC@172.17.6.10',6666).
It is also now corrected in the readme file.
