### 2022-01-11 eferrea/BCI-Master#3 @eferrea
Thanks also for pointing this out. I removed from the figure legend the word "extended" since the description is about all functionalities. I also included the full reference at the end of the readme.
