### 2022-01-11 eferrea/BCI-Master#4 @eferrea
Thanks that is right. I now changed the text to be a bit more clear:
After collecting enough samples by continuously reaching the target and updating by pressing the Update Regression Button (samples are shown on the right table, 150-200 for good performance), press the Switch BCI button to start the decoder.
