+++ paper/paper.md
@@ -48,5 +48,11 @@ The implemented decoder is a Kalman-Filter for motor control BCI applications [@
 
 ![Figure 1: BCI framework schematics. The BCImat framework interfaces with a task controller to display movements of a cursor on a computer screen and with a neural interface providing spiking signal to ultimately (after decoder calibration) control cursor movements. The neural signal fills an internal spike buffer. The spike buffer can be interfaced with an artificial spiking neural network (SNN) for stand-alone (patient independent) experiments (simulation mode), as well as with an external physiological intracortical interface providing spike data in real-time (application mode). Our system was tested in application mode with a Matlab executable (mex) code to stream online spikes recorded with a Cereplex system (Blackrock, Salt Lake City, USA). Task controller and BCI framework exchange messages, cursor position or velocities data via VRPN clients and servers. Inside the BCI framework, the cursor velocities are used to calibrate the decoder by regressing them with simultaneously recorded neural activity.  A task state parser is used internally to the framework to handle received messages from the BCI. It can be expanded to handle any type of message. Here for example, movement stage information and position of the target are handled to clear the spike buffer regularly to avoid excessive memory overload.](Fig1.png)
 
+# Funding:
+German Research Foundation (DFG, Germany, grant number FOR-1847-GA1475-B2). Received by AG.
+German Research Foundation (DFG, Germany, grant number SFB-889). Received by AG.
+Federal Ministry for Education and Research (BMBF, Germany, grant number 01GQ1005C). Received by AG.
+
+
 # References
 
