### 2022-01-11 #3956 @eferrea
@mstimberg ,  thanks for your comments. I answered your questions (eferrea/BCI-Master#2, eferrea/BCI-Master#3, eferrea/BCI-Master#4).
In addition, I have to mention that I just added in the paper.md a small paragraph of our funding sources.
As you suggested, I will then wait until the end of the review process to make an official release.
