+++ README.md
@@ -139,14 +139,17 @@ the same number of units should be maintained. Useful for now to restore previou
 
 2. Inside Test_connections:
  * TestVRPNConnection.m: is used to test if a VRPN client connection can be established without implementing all the BCI closed-loop.
- * Test_Blackrock: is used to test if a connection can be established with the Cereplex system.
+ * Test_Blackrock: is used to test if a connection can be established with the Cereplex system.  
+
+
+Have fun!
 
   
 **References**    
 WU, W., GAO, Y., BIENENSTOCK, E., DONOGHUE, J. P. & BLACK, M. J. 2006. Bayesian population decoding of motor cortical activity using a Kalman filter. Neural computation, 18, 80-118.
 
   
-  Have fun!
+
 
 
 
