### 2022-01-12 eferrea/BCI-Master#2 @mstimberg
> Yes, you are right. It should be:
> BCI_Loop(false,60,0.05,0,'TrackerBCI@172.17.6.10','TrackerTC@172.17.6.10',6666).
> It is also now corrected in the readme file.

I think the first (`isBrain=false`) is correct, but the second one with `isBrain=true` still mentions `TrackerBCI` twice.
