### 2022-01-13 #3956 @mstimberg
Thanks for the additional changes @eferrea. I just noticed that the paper still has a broken reference (`[RN1]` in line 76), but apart from that I am happy with the current state of the software/paper.
