+++ paper/paper.md
@@ -42,7 +42,7 @@ The code is intended for use by anyone wanting to test closed-loop BCI methods o
 The BCI framework (Fig.1) interfaces bidirectionally with a simple behavioral task controller (here written in c++) allowing to perform center-out reach movements for decoder calibration and then to control cursor movements via the neural activity. The communication between the task-controller and the BCI framework is done via the Virtual-Reality Peripheral Network (VRPN) protocol, implementing a client-server application via TCP or UDP [@RN1260] on both sides. On the task controller side, a standard c++ client-server application is used. On the BCImat side, a MATLAB executable version of the server and client VRPN classes are implemented to read the parameters for calibration and send the decoded parameters. Since the communication is established via IP network, the task controller and the BCI can run on different computers. 
 The framework is implemented in object-oriented MATLAB to exploit modularity. This makes it possible to interchange different decoders or to optimize decoding performance or to provide additional functionalities, for example, to perturb neural parameters for decoding in BCI learning experiments. Further supporting modularity, BCImat can communicate with other task controllers written in any other programming language provided that VRPN is used to stream and read information to and from the BCI framework. In addition, BCImat can be interfaced with any other recording hardware provided that spiking activity can be streamed in real-time and the Trial Spike Buffer format maintained.
 To use BCImat without recording hardware, an artificial spiking neural network (SNN) is implemented. The task-controller provided here allows performing a center-out reach task with computer mouse movements. While the subject performs the manual task, the simulated neurons fire accordingly with a Poisson process the frequency of which is proportional to the cosine of the angle between the direction of movements in the task and their preferred direction. Thereby, the SNN simulates neural response patterns of primate motor cortex during reaching tasks [@RN1261]. In practice, the user would first execute the manual motor task to calibrate the parameters of the decoding algorithm. Once switching from manual task execution to “BCI control” after calibration, the decoder output, i.e., the computer cursor movements, determines the neurons firing pattern depending on their own dynamics (cosine model) relative to the actual movement direction. During this closed-loop online control in the simulation mode, the cursor movements are therefore controlled via the decoder by the SNN and no longer by real computer mouse movements. Later, in the application mode, the subject’s brain activity would replace the SNN output. The code was successfully used from extracted neural activity of rhesus monkeys performing a similar center-out reach as previously described in a virtual-reality environment [@RN1049].  
-The implemented decoder is a Kalman-Filter for motor control BCI applications [@RN1070].   We also implemented some important BCI features that are frequently used in the literature to provide efficient training and better decoding performance. In particular, we implemented the possibility to re-train the Kalman filter during online control \cite{RN1}, an assisted computer cursor control during closed-loop trials [@RN589] to perform calibration in absence of movements, rotation of unit preferred directions resulting in movement direction rotations [@RN849], and the possibility to perform open-loop testing of the decoder (for review see [@RN61]).
+The implemented decoder is a Kalman-Filter for motor control BCI applications [@RN1070].   We also implemented some important BCI features that are frequently used in the literature to provide efficient training and better decoding performance. In particular, we implemented the possibility to re-train the Kalman filter during online control [@RN1], an assisted computer cursor control during closed-loop trials [@RN589] to perform calibration in absence of movements, rotation of unit preferred directions resulting in movement direction rotations [@RN849], and the possibility to perform open-loop testing of the decoder (for review see [@RN61]).
 
 #
 
