### 2022-01-21 #3956 @puolival
@oliviaguest It seems that I am not able to edit the checklist even though I am logged in.
