+++ README.md
@@ -156,6 +156,9 @@ the same number of units should be maintained. Useful for now to restore previou
  In this case you should be able to read the position of the cursor on the task controller screen.
  
  * Test_Blackrock: is used to test if a connection can be established with the Cereplex system.  
+this code snipped is based on how to read online data from cereplex system
+https://blackrockneurotech.com/research/wp-content/ifu/LB-0590-3.00-cbMEX-IFU.pdf
+the code reads spikes data for 5 seconds and restructure the data as used inside BCI mat. Importantly you must check that if spikes are recorded the buffer is not empty.
 
 
 Have fun!
