+++ BCImat/BCI_Classes/simNeurons_2D_velocity.m
@@ -6,7 +6,7 @@
 %At the actual implementation status, this class takes only the number of units as
 %input. 
 
-classdef simNeurons_2D_velocity < handle  % class should be implemented as a handle class. 
+classdef SimNeurons_2D_velocity < handle  % class should be implemented as a handle class. 
     % All copies of a given handle object refer to the same data.
     
     properties (Access=protected)
@@ -28,7 +28,7 @@ classdef simNeurons_2D_velocity < handle  % class should be implemented as a han
     %%%%% Methods
     methods (Access=public)
         %%%  constructor function
-        function obj = simNeurons_2D_velocity(N,varargin) %constructor for the class
+        function obj = SimNeurons_2D_velocity(N,varargin) %constructor for the class
             % protected memeber
             obj.time = tic;
             obj.start_time = tic;
