+++ README.md
@@ -28,7 +28,7 @@ SET(SFML_INCLUDE_PATH *\<change to the full path of your SFML include directory\
 SET(VRPN_INCLUDE_PATH *\<change to the full path of your VRPN include directory\>*)  
 SET(SFML_LIBRARY_PATH *\<change to the full path of your SFML lib directory\>*)  
 SET(VRPN_LIBRARY_PATH *\<change to the full path containing vrpn.lib\>*)  
-SET(VRPN_QUAT_LIBRARY_PATH *\<change to the full path containing quat.lib\>*)
+SET(VRPN_QUAT_LIBRARY_PATH *\<change to the full path containing quat.lib\>*)   
 SET(SFML_DLL_PATH *\<change to SFML folder containig DLLs, they should be in sfml-install-path/bin)
  
 In Linux, it is not necessary to specify these folders unless libraries are installed in custom locations.
