+++ BCImat/Test_connections/TestVRPNConnection.m
@@ -1,18 +1,18 @@
-%Check if VRPN messages and cursor positions are transmitted
+%% Check if VRPN messages and cursor positions are transmitted
 %@ E.Ferrea, 2015
 function TestVRPNConnection(server_address,ismessage)
-% input arguments:
+
+%% INPUT:
 %server_address: a vrpn server address from where to read the data;
 %ismessage: if =1 display messages read from the server, if =0
 %display curosr position
 
-
+%%
 
 fh = figure('Position',[1300 800 400 300])
 h = uicontrol('Style', 'PushButton', 'String', 'Stop BCI', ...
     'Callback', 'delete(gcbo)','Position',[100 100 200 100]);
 
-%server_address = 'TrackerTC@127.0.0.1:6666';%local computer
 vrpn_client('open_connection',server_address)
 
 
@@ -20,10 +20,10 @@ vrpn_client('open_connection',server_address)
 while ishandle(h)
     
     pause(0.02)
-    
+    %read messages and positions from server
     message= vrpn_client('get_messages',server_address);
     position =vrpn_client('get_positions',server_address);
-    
+    %display messages or positions
     if(ismessage)
         display(message)
     else
