+++ paper/paper.md
@@ -45,9 +45,10 @@ To use BCImat without recording hardware, an artificial spiking neural network (
 The implemented decoder is a Kalman-Filter for motor control BCI applications [@RN1070].   We also implemented some important BCI features that are frequently used in the literature to provide efficient training and better decoding performance. In particular, we implemented the possibility to re-train the Kalman filter during online control [@RN1], an assisted computer cursor control during closed-loop trials [@RN589] to perform calibration in absence of movements, rotation of unit preferred directions resulting in movement direction rotations [@RN849], and the possibility to perform open-loop testing of the decoder (for review see [@RN61]).
 
 #
+![Fig1](https://user-images.githubusercontent.com/40661882/151955047-94e8db54-1b68-4dae-9578-13f7f7c14a93.jpg)
 
 ![Figure 1: BCI framework schematics. The BCImat framework interfaces with a task controller to display movements of a cursor on a computer screen and with a neural interface providing spiking signal to ultimately (after decoder calibration) control cursor movements. The neural signal fills an internal spike buffer. The spike buffer can be interfaced with an artificial spiking neural network (SNN) for stand-alone (patient independent) experiments (simulation mode), as well as with an external physiological intracortical interface providing spike data in real-time (application mode). Our system was tested in application mode with a Matlab executable (mex) code to stream online spikes recorded with a Cereplex system (Blackrock, Salt Lake City, USA). Task controller and BCI framework exchange messages, cursor position or velocities data via VRPN clients and servers. Inside the BCI framework, the cursor velocities are used to calibrate the decoder by regressing them with simultaneously recorded neural activity.  A task state parser is used internally to the framework to handle received messages from the BCI. It can be expanded to handle any type of message. Here for example, movement stage information and position of the target are handled to clear the spike buffer regularly to avoid excessive memory overload.
-![Fig1](https://user-images.githubusercontent.com/40661882/151955047-94e8db54-1b68-4dae-9578-13f7f7c14a93.jpg)
+
 
 # Funding:
 German Research Foundation (DFG, Germany, grant number FOR-1847-GA1475-B2). Received by AG.
