+++ README.md
@@ -141,7 +141,7 @@ the same number of units should be maintained. Useful for now to restore previou
  
  This function needs to be called with the following arguments:
  
- TestVRPNConnection(server_address,ismessage).
+ test_VRPN_connection(server_address,ismessage).
  
  *server_address* requires the name of the server where the task controller runs, whereas *ismessage* specifies whether we want to read a message from the task controller (ismessage =1) or positional data from the task controller (ismessage =0).  
  
