### 2022-02-03 912c08c
JOSS review2
added comments, style
