+++ BCImat/Test_connections/test_VRPN_connection.m
@@ -1,8 +1,9 @@
-%% Check if VRPN messages and cursor positions are transmitted
-%@ E.Ferrea, 2015
-function TestVRPNConnection(server_address,ismessage)
 
-%% INPUT:
+function test_VRPN_connection(server_address,ismessage)
+% Check if VRPN messages and cursor positions are transmitted to BCImat
+%@ E.Ferrea, 2015
+
+% INPUT:
 %server_address: a vrpn server address from where to read the data;
 %ismessage: if =1 display messages read from the server, if =0
 %display curosr position
@@ -40,4 +41,3 @@ if ishandle(fh)
 end
 
 end
-    
