+++ BCImat/Test_connections/TestBlackrock.m
@@ -1,6 +1,7 @@
 %test connection with Blackrock cereplex system
-%Check that spike_data get filled during signal acquisition  
+%Check that spike_data get filled during signal acquisition
 %@ E. Ferrea,  2015
+
 %this code snipped is based on how to read online data from cereplex system
 %https://blackrockneurotech.com/research/wp-content/ifu/LB-0590-3.00-cbMEX-IFU.pdf
 %the code reads spikes data for 5 seconds and restructure the data as used
@@ -14,10 +15,10 @@ pause(5)
 event_data = cbmex('trialdata',1)
 cbmex('close')
 
- temp =  cellfun(@length,event_data(1:128,2:7));%to fit 3D Poisson spike class
-    
-    temp1 = temp(:)'; %reshape(A',1,[]);%
-    
-    spike_data(1,1:length(temp1)) = temp1;
-    
-    BCImat_format = reshape(spike_data,128,6);
+temp =  cellfun(@length,event_data(1:128,2:7));%to fit 3D Poisson spike class
+
+temp1 = temp(:)'; %reshape(A',1,[]);%
+
+spike_data(1,1:length(temp1)) = temp1;
+
+BCImat_format = reshape(spike_data,128,6);
