+++ BCImat/GUI_functions/check_correlation.m
@@ -1,0 +1,13 @@
+function check_correlation (object, eventdata)
+% to display correlation values between real and decoded movemnets in the idle mode
+% @E.Ferrea,2015
+
+%INPUT
+%take decoder object as input and display check Pearson Correlation
+%coefficients
+
+%%
+object.check_correlation = true;
+disp('Check Correlation')
+
+end

[…]
