+++ paper/paper.bib
@@ -68,7 +68,7 @@
 @article{RN1,
    author = {Gilja, V. and Nuyujukian, P. and Chestek, C. A. and Cunningham, J. P. and Yu, B. M. and Fan, J. M. and Churchland, M. M. and Kaufman, M. T. and Kao, J. C. and Ryu, S. I. and Shenoy, K. V.},
    title = {A high-performance neural prosthesis enabled by control algorithm design},
-   journal = {Nat Neurosci},
+   journal = {Nature Neuroscience},
    volume = {15},
    number = {12},
    pages = {1752-7},
@@ -98,7 +98,7 @@
 @article{RN849,
    author = {Jarosiewicz, B. and Chase, S. M. and Fraser, G. W. and Velliste, M. and Kass, R. E. and Schwartz, A. B.},
    title = {Functional network reorganization during learning in a brain-computer interface paradigm},
-   journal = {Proc Natl Acad Sci U S A},
+   journal = {The Proceedings of the National Academy of Sciences},
    volume = {105},
    number = {49},
    pages = {19486-91},
