+++ paper/paper.md
@@ -32,23 +32,26 @@ bibliography: paper.bib
 ---
 
 # Summary
-Recent advances in intracortical Brain-Computer Interface (BCI) technology allowed motor disabled patients to partially regain lost motor functions [@RN572;@RN1272;@RN589;@RN569]. In these patients, intact neural activity is extracted from motor-related areas of the cerebral cortex via intracortical implanted electrodes and interpreted by a machine-learning algorithm to control a prosthetic device, thereby bypassing dysfunctional corticospinal projections that resulted, for example, from spinal cord lesions.  BCIs are also used for basic neuroscientific studies to establish a specific transformation between the brain area under investigation and a specific behavior[@RN877;@RN219] and therefore imposing a direct causal link between them. The software introduced here allows true online BCI control of a computer cursor based on physiological signals (e.g. from patients) as well as realistic real-time simulations for testing all algorithms based on artificial spiking neural network (SNN) data using the identical control architecture.
+Recent advances in intracortical Brain-Computer Interface (BCI) technology allowed motor disabled patients to partially regain lost motor functions [@RN572;@RN1272;@RN589;@RN569]. In these patients, intact neural activity is extracted from motor-related areas of the cerebral cortex via intracortical implanted electrodes and interpreted by a machine-learning algorithm to control a prosthetic device, thereby bypassing dysfunctional corticospinal projections that resulted, for example, from spinal cord lesions.  BCIs of this type have been and still are being developed mostly in non-human primate animal models. Additionally, they are also used for basic neuroscientific studies in animals to establish a specific experimentally-controlled transformation between the brain area under investigation and a specific behavior[@RN877;@RN219], thereby imposing a direct and controllable causal link between brain activity and behavior. The software introduced here allows true online BCI control of a computer cursor based on physiological signals. Importantly, it also allows realistic real-time neural data simulations from artificial spiking neural network (SNN). With this, all algorithms and the control architecture can be tested in silico identical to the physiological experiment.
 
 # Statement of need
-Most of the publicly available software for BCIs is designed for applications based on time-continuous electrophysiological signals, like electroencephalographic (EEG) or electrocorticographic (ECoG) signals [@RN1263] and requires existing recording possibilities or pre-recorded data for testing. BCImat instead is a MATLAB framework for implementing and testing a BCI which is based on stochastic event time-series, particularly neuronal spiking signals recorded from large numbers of individual neurons that vary at a time-scale of milliseconds. For example, BCImat can use intracortical single-neuron signals acquired with a Cereplex system (Blackrock, Salt Lake City, USA). Importantly, BCImat can alternatively use the input from a built-in artificial spiking neural network (SNN) for simulating the later online BCI control. This allows testing BCI applications with the same algorithm and framework as intended for later use in patients, but before the availability of implanted subjects and specific recording hardware. This way, the full online decoding experiment or application can be developed in advance without the need for pre-recorded data files. 
-The code is intended for use by anyone wanting to test closed-loop BCI methods or perform intracortical closed-loop BCI experiments.
[…]
