+++ paper/paper.md
@@ -51,7 +51,7 @@ The implemented decoder is a Kalman-Filter for motor control BCI applications [@
 
 
 
-![Figure 1: BCImat framework. The BCImat framework interfaces with a task controller to display movements of a cursor on a computer screen and with a neural interface providing spiking signal to ultimately (after decoder calibration) control cursor movements. The neural signal fills an internal spike buffer. The spike buffer is fed with input from an artificial spiking neural network (SNN) for stand-alone (patient independent) experiments (simulation mode), or with external intracortical physiological signals providing neural spike data in real-time (application mode). Our system was tested in application mode with a Matlab executable (mex) code to stream online spikes recorded with a Cereplex system (Blackrock, Salt Lake City, USA). Task controller and BCI framework exchange messages, cursor position or velocities data via VRPN clients and servers. Inside the BCI framework, the cursor velocities are used to calibrate the decoder by regressing them with simultaneously recorded neural activity.  A task state parser is used in BCImat to handle received messages from the task controller. It can be expanded to handle any type of message. Here, for example, information about the completion of a movement is handled to clear the spike buffer regularly to avoid memory overload.](https://user-images.githubusercontent.com/40661882/153038621-e3cafe89-714f-4538-9295-e8a6a48db9fa.jpg)
+![BCImat framework. The BCImat framework interfaces with a task controller to display movements of a cursor on a computer screen and with a neural interface providing spiking signal to ultimately (after decoder calibration) control cursor movements. The neural signal fills an internal spike buffer. The spike buffer is fed with input from an artificial spiking neural network (SNN) for stand-alone (patient independent) experiments (simulation mode), or with external intracortical physiological signals providing neural spike data in real-time (application mode). Our system was tested in application mode with a Matlab executable (mex) code to stream online spikes recorded with a Cereplex system (Blackrock, Salt Lake City, USA). Task controller and BCI framework exchange messages, cursor position or velocities data via VRPN clients and servers. Inside the BCI framework, the cursor velocities are used to calibrate the decoder by regressing them with simultaneously recorded neural activity.  A task state parser is used in BCImat to handle received messages from the task controller. It can be expanded to handle any type of message. Here, for example, information about the completion of a movement is handled to clear the spike buffer regularly to avoid memory overload.](https://user-images.githubusercontent.com/40661882/153038621-e3cafe89-714f-4538-9295-e8a6a48db9fa.jpg)
 
 # Funding:
 German Research Foundation (DFG, Germany, grant number FOR-1847-GA1475-B2). Received by AG.
