### 2022-02-21 #3956 @mstimberg
To get the rendered PDF here we should also be able to use
