### 2022-03-09 #3956 @mstimberg
Hi everyone. In case there are any doubts about this: I still recommend the acceptance of the paper in its most recent version.
Quite obviously @puolival's review was  more "demanding" than my own, but I also agree that the changes to address the raised points improved the manuscript/code/documentation.

I hope I am not overstepping my role as a reviewer here, but let me mention two points from the JOSS guidelines that might affect the evaluation of the current manuscript and code:

> Your paper must not focus on new research results accomplished with the software.

([Submission guidelines for authors](https://joss.readthedocs.io/en/latest/submitting.html)) I think it is important to keep in mind that the role of papers in JOSS is different from a "software paper" in most other journals.

> Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.
>
>    Good: An automated test suite hooked up to continuous integration (GitHub Actions, Circle CI, or similar)
>    **OK: Documented manual steps that can be followed to objectively check the expected functionality of the software (e.g., a sample input file to assert behavior)**

([Review criteria: tests](https://joss.readthedocs.io/en/latest/review_criteria.html#tests)) While I of course agree that unit tests are important, I also acknowledge that the core part of the presented work involves GUI and client-server communication, which means that a fully automated test suite presents a considerable effort.
