### 2022-03-17 #3956 @puolival
@oliviaguest @eferrea

Thank you for the revised manuscript and software. Overall, I think this software has great potential to become a very useful tool for the BCI community. However, while some improvements have been now made after the initial submission, many of my previous major comments have not been addressed. Therefore, I cannot recommend accepting this submission at this stage.

I have here summarized part of the issues that I already pointed out in the previous review:

1. The manuscript is too underdeveloped. There is no full description of the performed experiments or the obtained results. For example, the manuscript does not contain data that supports the claim "To demonstrate functionality and performance, the BCImat code was successfully used with neural activity recorded in motor cortical areas of two rhesus monkeys performing a center-out reach task in a virtual-reality environment, .." What exactly were those experiments and how were they conducted? What were the results? Since no complete data from an actual experiment has been reported, it is impossible to tell whether the software and methods actually work or not. Information about the ethical approvals has been added, which is good.

2. Evidence for the translational value must be added to the manuscript. If this software is intended to be used in recordings done on humans, there must be data that shows the software works in such context.

3. There are no comparisons to other BCI software.

4. There is still no automated package management solution and installing this software still requires substantial effort on the end user's part: locating, downloading, and installing external libraries, compiling the task controller, and editing environmental variables and configuration files. Ideally, there would be just one installation script that the user needs to run, or an equivalent simple solution. Compare the situation for example to installing Python software libraries, where something like "pip install bcimat" could be sufficient in an ideal case.

5. Automated tests have not been added. I agree with @mstimberg that writing some of these tests could require considerable effort but this is a JOSS requirement. Also, even limited testing would be much better than no testing at all.

6. My previous comment "Many functions and methods are very long, might do several things, and are often not easy to understand. I suggest that long functions are refactored into several smaller ones, such that each function does one only thing" has not been completely addressed. Input/output variables remain uncommented in many parts. Like previously, it is still on many occassions not possible to understand what the variables are for, what are the allowed data/value ranges, etc. Some variables have been renamed and comments added but for the most part the software has not been revised for greater clarity.

7. The drawbacks of having no connectivity between the neurons in the SNN should be acknowledged. This makes the simulation biophysically very unrealistic and there probably can't be population-level phenomena typical to real recordings, e.g. oscillations in the population activity. The claim "Importantly, it also allows realistic real-time neural data simulations .." in the abstract is now too strong. No additional GUI was implemented for adjusting the parameters of the SNN.

8. Please add the required version numbers also for MATLAB and the Statistics and Machine Learning Toolbox.

Finally, I want to emphasize that all comments are intended as constructive criticism. In my opinion, addressing the issues pointed out here and in the previous review would make it much easier to adopt this software into one's use and/or help contribute to the software.
