### 2022-03-17 #3956 @mstimberg
My earlier comment probably made clear that I do not quite share @puolival's assessment, in particular that the "manuscript is underdeveloped" – as @eferrea pointed out, the manuscript is already longer than recommended. But of course my opinion shouldn't have any more weight here than anyone else's.

I just wanted to comment on one point that referred to my earlier comment:

> 5\. Automated tests have not been added. I agree with @mstimberg that writing some of these tests could require considerable effort but this is a JOSS requirement. Also, even limited testing would be much better than no testing at all.

Automated tests are "strongly encouraged" by JOSS, but they are not a requirement. The [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html) explicitly state that "Documented manual steps that can be followed to objectively check the expected functionality" are considered "OK" .
