### 2022-03-18 #3956 @puolival
@mstimberg Thank you for clarifying this. If my reading of the guidelines was too strict, I am happy to continue as you see best @oliviaguest.
