### 2022-03-21 #3956 @oliviaguest
Thanks, all.

I'm going to think about some points raised above, but I do think 4, 6, 8, pasted below, are worthy of comment/consideration from the authors:

> 4. There is still no automated package management solution and installing this software still requires substantial effort on the end user's part: locating, downloading, and installing external libraries, compiling the task controller, and editing environmental variables and configuration files. Ideally, there would be just one installation script that the user needs to run, or an equivalent simple solution. Compare the situation for example to installing Python software libraries, where something like "pip install bcimat" could be sufficient in an ideal case.

> 6. My previous comment "Many functions and methods are very long, might do several things, and are often not easy to understand. I suggest that long functions are refactored into several smaller ones, such that each function does one only thing" has not been completely addressed. Input/output variables remain uncommented in many parts. Like previously, it is still on many occassions not possible to understand what the variables are for, what are the allowed data/value ranges, etc. Some variables have been renamed and comments added but for the most part the software has not been revised for greater clarity.

> 8. Please add the required version numbers also for MATLAB and the Statistics and Machine Learning Toolbox.
