### 2022-03-22 #3956 @eferrea
Dear Editor (@oliviaguest), many thanks for restricting the review to the most reasonable points. We see the benefit in addressing point 6, which can be achieved with reasonable effort given that the code is also already shared inside the lab and with external collaborators. Point 8 is very minor and we have already addressed it in the readme file. Regarding point 4, we were wondering whether in its actual version installation instructions are not already satisfying JOSS guidelines as we read them at https://joss.readthedocs.io/en/latest/review_checklist.html since automated package management solution is not a strict requirement and a list of dependencies is clearly-stated:

* Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

We appreciate your time and effort in considering our manuscript and look forward to your feedback.
Kind regards, Enrico Ferrea (on behalf of all authors)
