### 2022-03-25 8d1f8b9
JOSS review
add missing input/output comments
