+++ BCImat/display/display_table.m
@@ -1,10 +1,17 @@
-
-function display_table(f,position_x,position_y, length_x,length_y,decoder,calibrator)
 % this function display the list of parameters in a table in a GUI,
 %@ E.Ferrea, 2015
+function display_table(f,position_x,position_y, length_x,length_y,decoder,calibrator)
 
-%INPUT 
-%It takes the position and length of the table as input and decoder and calibration objects 
+
+%INPUT:
+
+%It takes the position and length of the table, decoder and calibration objects as input.
+%position_x: horizintal position of the window.
+%position_y: vertical position of the window.
+%length_x: window length in the horizontal dimension.
+%length_y: window length in the vertical dimension.
+%decoder: name of the decoder object.
+%calibrator: name of the calibrator object.
 
 
 [row,col] = find(decoder.neurons > 0);
