+++ BCImat/BCI_Classes/Kalman_decoder_class.m
@@ -1,13 +1,13 @@
 % Decoder class for Kalman Filter decoding approach as described in Wu et al., 2006.
 % Decoder position is updated only during stage number 2. This value in
 % hard coded here but should be ideally placed in the constructor.
-% We decided to update the velocity values only during stage 2 (line 128 in loop function) as for
+% We decided to update the velocity values only during stage 2 as for
 % some standard BCI control experiments.
 
 %@E.Ferrea, 2015
 
 classdef Kalman_decoder_class < handle;
-    properties (Access=private)  
+    properties (Access=private)
         K;
         iteration_step;
         X_prior;
@@ -42,7 +42,7 @@ classdef Kalman_decoder_class < handle;
         neurons;
         H;
         H_reset;
-        fixation_position;   
+        fixation_position;
     end
     
     properties (SetAccess = public)
@@ -51,14 +51,23 @@ classdef Kalman_decoder_class < handle;
         n_neurons;
         X_update;
         sample;
-        check_correlation;     
+        check_correlation;
     end
     
     methods (Access=public)
         
         %Constructor class -> initialize parameters used for decoding
-        function obj = Kalman_decoder_class(sample_size,max_exp_duration,delay,max_samples)       
-            obj.expected_total_samples = ceil(max_exp_duration/sample_size);
+        function obj = Kalman_decoder_class(BCI_update_time,max_exp_duration,delay,max_corr_samples)
+            
+            %INPUT
+            %BCI_update_time: choose in seconds the rate of the BCI. (0.05 s is suggested).
+            %max_exp_duration: maximum expected time of experiment duration in seconds.
+            %delay: number of time bins (!!! not ms) of BCI_update_times to shift the neural activty relative to motor output. This can be used in real experiments to make up for delays
+            %in transmission of the neural signal through the spinal cord. Phisiological values are in the range of 100-150 ms). For testing purposes use 0 (0 ms no delay)
+            %or 1 (1*BCI_update_time ms).
+            %max_corr_samples: specify the max number of samples allowed for estimating a correlation during IDLE contol.
+            
+            obj.expected_total_samples = ceil(max_exp_duration/BCI_update_time);
             obj.K = 0;
             obj.position = zeros(obj.expected_total_samples,2);
             obj.velocity = zeros(obj.expected_total_samples,2);
@@ -74,11 +83,11 @@ classdef Kalman_decoder_class < handle;
             obj.firing_rate = zeros(obj.expected_total_samples,768);
             obj.sample = obj.delay;
             obj.correlation_counter  = 0;
-            obj.max_corr_samples = max_samples;
+            obj.max_corr_samples = max_corr_samples;
             obj.correlation_velocity = zeros(3,obj.max_corr_samples);
             obj.bci_correlation_velocity = zeros(3,obj.max_corr_samples);
             obj.check_correlation = false;
-            obj.dt = sample_size;
+            obj.dt = BCI_update_time;
             obj.rotation_counter = 1;
             obj.fixation_position = [0 0];
             
@@ -92,13 +101,15 @@ classdef Kalman_decoder_class < handle;
             obj.fileID_decoder = fopen(filename_decoder,'at');
             obj.printfFormatHeader = ['%6s' '%6s\n'];
             obj.printfFormatBody = ['%12.3f  %12.3f\n'];
-            fprintf(obj.fileID_decoder,obj.printfFormatHeader,'time','K');          
+            fprintf(obj.fileID_decoder,obj.printfFormatHeader,'time','K');
         end
         
         %set BCI starting postion from outside (this info comes from the
         %task controller (usually). It is used to place the cursor at the
         %central fixation target when control is not required.
         function obj = set_fixation_position(obj,fixation_position)
+            %INPUT:
[…]
