### 2022-03-25 #3956 @eferrea
@puolival , @oliviaguest , @mstimberg

Thanks for taking the time to review our software. Here we address/discuss at our best the remaining points.

4. There is still no automated package management solution and installing this software still requires substantial effort on the end user's part: locating, downloading, and installing external libraries, compiling the task controller, and editing environmental variables and configuration files. Ideally, there would be just one installation script that the user needs to run, or an equivalent simple solution. Compare the situation for example to installing Python software libraries, where something like "pip install bcimat" could be sufficient in an ideal case.

So far, we implemented @mstimberg as well as @puolival suggestions to make the installation procedure simpler. We think that after these improvements, the actual version provides a clearly stated list of dependencies that can be followed to install the software which is the specific JOSS requirement.

6. My previous comment "Many functions and methods are very long, might do several things, and are often not easy to understand. I suggest that long functions are refactored into several smaller ones, such that each function does one only thing" has not been completely addressed. Input/output variables remain uncommented in many parts. Like previously, it is still on many occasions not possible to understand what the variables are for, what are the allowed data/value ranges, etc. Some variables have been renamed and comments added but for the most part, the software has not been revised for greater clarity.

We now add missing comments documenting all input/output variables to improve understanding of what these variables do inside the functions. We specified which data ranges are used in the calibrator function (see BCI_loop.m lines 102-103). We believe that in the current version the code can be more easily read. However, we did not find excessively long functions that could be refactored into smaller ones without impacting the core of the project that is already shared inside the lab and with external collaborators.

8. Please add the required version numbers also for MATLAB and the Statistics and Machine Learning Toolbox.

Used Matlab version (2016a) and toolbox version (10.2) are now specified in the readme file.

We hope that after considering the points raised by the reviewer, the manuscript now fulfills the criteria for publication.
