### 2022-03-29 9965ca6
Upgrade functionalities
Port should not be necessarily specified (as could be the case on two different computers);
Implemented close txt file for decoder
