+++ README.md
@@ -133,7 +133,7 @@ the same number of units should be maintained. Useful for now to restore previou
 
  * SimNeurons_2D_velocity.m: this class is used inside BCI_loop.m to generate firing rates of artificial neurons according to a Poisson distribution. During each small movement (e.g 50 ms timestamp), the mean of each Poisson neuron is determined by the baseline firing rate plus the cosine of the actual moving direction and the neuron preferred direction multiplied by the modulation depth. Each neuron preferred direction is determined according to a circular uniform distribution.  This class also uses a hardcoded modulation depth in the physiological range of 4-18 Hz and a baseline firing rate in the range 4-20.  Future releases of this class could include the possibility of specifying these ranges. 
 
- * task_parser.m: this simple class is used to handle messages coming from the task controller. It is important for the BCImat to know information about the state of the task    to adapt its behavior accordingly. For example, it could be useful to collect samples of firing rates and positional samples only when the users are engaged with the motor task. In this case, we decided to store the values for calibration online during the last stage of the task (stage 3). 
+ * Task_parser.m: this simple class is used to handle messages coming from the task controller. It is important for the BCImat to know information about the state of the task    to adapt its behavior accordingly. For example, it could be useful to collect samples of firing rates and positional samples only when the users are engaged with the motor task. In this case, we decided to store the values for calibration online during the last stage of the task (stage 3). 
 
  * Kalman_calibrator_class.m: this class is used to internally store samples for calibrating a BCI decoder according to the following paper (Wu et al. 2006).
  * Kalman_decoder_class.m: it uses the calibration matrix generated by the Kalman_calibration class to generate position estimation online. Ideally, this class and the calibrator class could be replaced with any other decoder type implementing a similar behavior. 
