### 2022-03-30 #3956 @puolival
@oliviaguest

I am still not able to modify the list. If I click the previous invitation link again, I get the following error message "Sorry, we couldn't find that repository invitation. It is possible that the invitation was revoked or that you are not logged into the invited account."

Anyway, as it was decided that a large part of my review will not be addressed, I think you can continue with the submission as you see best, and that further input wouldn't change this. Thanks.
