### 2022-03-30 #3956 @danielskatz
@puolival - because JOSS changed its bot, some reviews that were started but where a reviewer didn't accept an invitation ended up in an odd situation.  The best option here is for you to write `@editorialbot generate my checklist` as a new comment - this will generate a new checklist for you that you can then fill in.  You can then ignore your part of the original checklist (and @oliviaguest could edit it to remove your part if she wished)
