### 2022-04-06 #3956 @oliviaguest
@eferrea and @puolival I did not restrict the review. Let's be very clear here, what I said above in https://github.com/openjournals/joss-reviews/issues/3956#issuecomment-1073573426 was this:

> I'm going to think about some points raised above, but I do think 4, 6, 8, pasted below, are worthy of comment/consideration from the authors:

Saying something needs to be thought about while highlighting some parts does not mean the review is restricted to only those parts. I am going to think further about this, but please do not misunderstand me.
