### 2022-04-11 #3956 @oliviaguest
@puolival may I ask you to separate your review to differentiate between 'essential' (must address) feedback and 'nice to have' from your perspective, please? And then, we can see what compromises can be made here as indeed it's not always possible for reviewers and authors to agree on what essential changes are, and it is my job to do so. I would like to underline that I am very thankful, and I am sure the authors are too, for the time you have invested here to give such detailed feedback and your care and attention is greatly appreciated.
