### 2022-05-11 #3956 @eferrea
Dear @oliviaguest, given the actual situation, we kindly ask you how to proceed from here. So far we addressed or discussed all the main remaining issues but unfortunately we did not received any feedback yet (see https://github.com/openjournals/joss-reviews/issues/3956#issuecomment-1078829728 and https://github.com/openjournals/joss-reviews/issues/3956#issuecomment-1070967659.
Because of that we think that in its current version ( and also as @mstimberg pointed out https://github.com/openjournals/joss-reviews/issues/3956#issuecomment-1063096197) the software is suitable for publication.
