### 2022-05-16 #3956 @eferrea
Hi @oliviaguest, many thanks for your answer . From our side is fine to be patient (we also were already :) ), however we would like to know whether @puolival is still willing to finish the review given that he did not replied any messages since before Easter. If he wants to continue we can wait, otherwise we think that would be better to take an alternative decision now. Maybe he could shortly comment on that.
