### 2022-05-31 #3956 @eferrea
Dear @oliviaguest, at this point we wanted to solicit a decision upon our manuscript. On March 30th reviewer 2 said “Anyway, […], I think you can continue with the submission as you see best, and that further input wouldn't change this. Thanks.”. From our side it is obvious that reviewer 2 clearly indicated with this statement that they have no intention to continue with the review; consequently, no further messages followed this statement since 2 months.  Further patience for “few weeks”, as you suggest, will almost certainly not change this situation.

From our side, we answered or discussed all the raised points in our previous reply https://github.com/openjournals/joss-reviews/issues/3956#issuecomment-1078829728. To summarize, we added comments to the code for better clarity and we modified the text to fulfill points 6,8 of comment https://github.com/openjournals/joss-reviews/issues/3956#issuecomment-1070967659. We discussed that points 4-5 are already satisfied according to journal guidelines, whereas  point 1  was already addressed in our previous revision by adding information on average target acquisition time as an effective measure of BCI performance. As already mentioned, we could add several and similar measurements but this will add additional text to the paper without adding additional important information. We also discuss that our software for intracortical BCI cannot really be compared with available EEG based software given the very different types of physiological signals. We think that at this stage, the main points from two reviewers are largely satisfied. We therefore would kindly ask you to accept the paper in its current version to avoid further delay since we are by now 9 months after first submission.
Enrico Ferrea on behalf of all authors
