### 2022-05-31 #3956 @danielskatz
👋 @eferrea and @oliviaguest - I'm writing as the JOSS Associate-Editor-in-Chief on duty this week.

I apologize to @eferrea for the length of time this is taking.

However, peer review does not succeed by time passing.  I think @oliviaguest needs to find another reviewer who can carry our a review quickly if possible, or perhaps step in and perform the review herself if she has time.
