### 2022-06-02 #3956 @danielskatz
@oliviaguest - Having re-read this thread, I think you can make a decision if you think that the comments from @puolival's review needed to make this acceptable to JOSS have been met, in the absence of further feedback from @puolival.

Alternatively, you can try to find a new reviewer.

If you think I've missed an option, please let me know.
