+++ paper/paper.bib
@@ -60,7 +60,7 @@
    number = {11},
    pages = {1527-1537},
    ISSN = {0270-6474},
-   DOI = {doi.org/10.1523/JNEUROSCI.02-11-01527.1982},
+   DOI = {10.1523/JNEUROSCI.02-11-01527.1982},
    year = {1982},
    type = {Journal Article}
 }
