+++ paper/paper.md
@@ -9,7 +9,7 @@ tags:
 
 authors:
   - name: Enrico Ferrea^[corresponding author]
-    affiliation: 1
+    affiliation: "1,5"
   - name: Pierre Morel
     affiliation: "1,4" 
   - name: Alexander Gail
@@ -22,8 +22,10 @@ affiliations:
    index: 2
  - name: Bernstein Center for Computational Neuroscience, Goettingen, Germany
    index: 3
- - name: Univ. Littoral Côte d’Opale, Univ. Lille, Univ. Artois, ULR 7369 - URePSSS - Unité de Recherche Pluridisciplinaire Sport Santé Société, F-59140 Dunkerque, France
+ - name: ULR 7369 ‑ URePSSS ‑ Unite de Recherche Pluridisciplinaire Sport Santé Sociéte, Univ. Littoral Côte d’Opale, Univ. Lille, Univ. Artois, Calais, France. 
    index: 4
+ - name: Institute for Neuromodulation and Neurotechnology, University Hospital and University of Tuebingen, Tuebingen, Germany
+   index: 5  
 
 date: 16 August 2021
 bibliography: paper.bib
