+++ paper/paper.bib
@@ -52,6 +52,18 @@
    type = {Journal Article}
 }
 
+@article{ferrea2022statistical,
+  title={Statistical determinants of visuomotor adaptation along different dimensions during naturalistic 3D reaches},
+  author={Ferrea, E and Franke, J and Morel, P and Gail, A},
+  journal={Scientific Reports},
+  DOI = {10.1038/s41598-022-13866-y},
+  volume={12},
+  number={1},
+  pages={1--14},
+  year={2022},
+  publisher={Nature Publishing Group}
+}
+
 @article{RN1261,
    author = {Georgopoulos, Apostolos P and Kalaska, John F and Caminiti, Roberto and Massey, Joe T},
    title = {On the relations between the direction of two-dimensional arm movements and cell discharge in primate motor cortex},
