+++ paper/paper.md
@@ -45,7 +45,7 @@ The software that defines the BCI framework (BCImat, Fig. 1) interfaces bi-direc
 The BCI framework is implemented in object-oriented Matlab to exploit modularity. This makes it possible to use more advanced decoders (e.g., non-linear long-short-term-memory (LSTM) networks, Transformers) as alternatives in order to optimize decoding performance or to provide additional functionalities (e.g., to perturb neural parameters for decoding to experimentally induce neural plasticity in BCI learning studies). Further supporting modularity, BCImat can communicate with other task controllers written in any other programming language provided that VRPN is used to stream and read information to and from the BCI framework. For real intracortical BCI experiments (application mode, Fig. 1), the presented package is implemented for interfacing a Cereplex system (Blackrock, Salt Lake City, USA), but can easily be adapted to other common data acquisition systems. In fact, BCImat can be interfaced with any other recording hardware provided that the same internal buffer structure to store spiking activity in real-time is programed.
 To use BCImat without recording hardware, an artificial spiking neural network (SNN) is implemented (simulation mode, Fig. 1). In the simulation mode, we decided to implement the same spike buffer structure of the application mode in order to keep compatibility when switching among the two modes. The task-controller provided here allows performing a center-out reach task with computer mouse movements. While the subject performs the task manually, i.e. by actually physically moving the arm (manual task), the simulated neurons fire accordingly. Mimicking stochastic properties of neural activity in the brain, the firing pattern is simulated as a Poisson process. The frequency of firing is proportional to the cosine of the angle between the direction of movements in the task and their preferred direction. Thereby, the SNN simulates neural response patterns of primate motor cortex during reaching tasks [@RN1261]. 
 In practice in the simulation mode, the user would first execute the manual task to calibrate the parameters of the decoding algorithm. After calibration, one will switch from manual task execution to closed-loop “BCI control”. During this closed-loop, the cursor movements are controlled via the decoder output and no longer by real computer mouse movements. At the same time, the user should still perform the task with the mouse, since the cursor movement determines the neurons firing pattern of the SNN, which serves as input to the decoder. The SNN firing depends on its own neural dynamics driven by the cosine model relative to the mouse movement direction. Later, in the application mode, the subject’s brain activity would replace the SNN output, and physical movements would be no longer required to move the cursor. 
-To demonstrate functionality and performance, the BCImat code was successfully used with neural activity recorded in motor cortical areas of two rhesus monkeys performing a center-out reach task in a virtual-reality environment, similar to a setup that was previously described [@RN1049].  Both animals were housed in social groups with one or two male conspecifics in facilities of the German Primate Center. The facilities provide cage sizes exceeding the requirements by German and European regulations, and access to an enriched environment including wooden structures and various toys. All procedures have been approved by the responsible regional government office [Niedersächsisches Landesamt für Verbraucherschutz und Lebensmittelsicherheit  (LAVES)] under permit numbers 3392 42502-04-13/1100 and comply with German Law and the  European Directive 2010/63/EU regulating use of animals in research. Both animals learned to control the cursor via BCImat. The time that the animals needed to move the cursor to the targets was in the same order of magnitude during manual and BCI task performance in both animals (animal 1: median hand movement time = 404 ms,  animal 1: median BCImat movement time = 585 ms, animal 2: median hand movement time = 418 ms,   animal 2: median BCImat movement time = 743 ms). Note that the observed decrease in performance comparing hand movements with BCI movements is generally expected.
[…]
