+++ paper/paper.bib
@@ -43,14 +43,6 @@
    type = {Journal Article}
 }
 
-@article{RN1049,
-   author = {Ferrea, Enrico and Morel, Pierre and Franke, Joscha and Gail, Alexander},
-   title = {Statistical determinants of visuomotor adaptation in a virtual reality three-dimensional environment},
-   journal = {bioRxiv},
-   DOI = {10.1101/2021.02.16.431440},
-   year = {2021},
-   type = {Journal Article}
-}
 
 @article{ferrea2022statistical,
   title={Statistical determinants of visuomotor adaptation along different dimensions during naturalistic 3D reaches},
