### 2022-06-27 #3956 @eferrea
@oliviaguest thanks for your message, I archived the software on Zenodo [DOI: 10.5281/zenodo.6759182] . I also updated the affiliations and updated bibliography with one paper that was on the bioRxiv and recently published.
