### 2022-07-04 #3956 @oliviaguest
@eferrea sorry, but why? Didn't their contribution get used? This is a strange request unless you think their labour was worth nothing to you.
