### 2022-07-04 #3956 @mstimberg
@oliviaguest (sorry, forgot to tag you earlier): to avoid misunderstandings, @eferrea's comment was related to my earlier comment:
> I certainly don't intend to complicate things and to draw out the publishing process further, but I wonder: other journals I've reviewed for that mention the reviewer names in the paper state something along the lines of "reviewers endorsed the manuscript in its current form", although I'm not sure if that's the case for JOSS? I'm saying this since I'm not 100% sure that @puolival would subscribe to that statement and might not be happy to be listed on the paper.

I am completely convinced that @puolival *deserves* to be listed to be a reviewer, I am just not 100% sure that they *want* to be listed as a reviewer.
