### 2022-07-04 #3956 @oliviaguest
I appreciate you raising this, but I don't appreciate being told, even if indirectly, how to manage this. I'll happily give @puolival time to reply, but I'm wondering if maybe an @openjournals/joss-eics should take over completely from here. I'm happy to back off, especially if my job is almost completely done anyway.
