### 2022-07-04 #3956 @danielskatz
Given that the entire review history is public, I don't think that we need additional statements from reviewers. This hasn't been the practice of JOSS in the past.
