### 2022-07-05 #3956 @danielskatz
The DOI doesn't yet resolve for me, so I'll keep this open until it does.
